split contentviews.py into mitmproxy.contentviews
